gtksourceview4 dependency check: Allowing match with "gtksourceview-4"
This changeset it to address a potential issue that may be encountered
when installing the Ruby-GNOME gtksourceview4 gem under bundler.

Synopsis:

The gtksourceview4 installation from FreeBSD ports will install a
pkg-config data file at the pathname
  /usr/local/libdata/pkgconfig/gtksourceview-4.pc

When the Ruby-GNOME gtksourceview4 gem is installed under bundler on a
FreeBSD host, then the dependency test for gtksourceview-4.0.pc
may fail due to the absence of such a filename for the *.pc file.

This may not have been encountered in packaging for the FreeBSD ports
for Ruby-GNOME, due to the differing installation procedure - such as
when installing these ports with 'gem', installed to a temporary staging
directory in the package creation process. A complete analysis of this
installation process would be beyond the scope of this changelog synopsis.

For some Bundle source configurations - such as when installing
Ruby-GNOME gems from a local Git working tree with Bundler - Bundler
will run the dependency-check/Rakefile during the gtksourceview4
gem installation. Thus, due to the mis-match in the name for the
expected *.pc file, the gem cannot be sucessfully installed with
Bundler, in these configurations.

Summary:

This changeset presents one approach towards addressing this concern of
the differing *.pc filename for the GNOME gtksourceview4 distribution on
some hosts.